refactor(types): create global.d.ts
